+++ src/main/java/com/example/authmodule/repository/VagaRepository.java
@@ -5,6 +5,6 @@ import com.example.authmodule.models.Vaga;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VagaRepository extends JpaRepository<Vaga, Long>{
+public interface VagaRepository extends JpaRepository<Vaga, Integer>{
 
 }
